Refactor: Changed structure of files BusinessSection ContactSection FeatureSection FooterSection
