+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 import BackgroundPattern from "@/components/Background";
-import BusinessOutcomes from "@/components/LandingPage/Business/business";
-import ContactSection from "@/components/LandingPage/ContactSection.tsx/ContactSection";
-import FeaturesPage from "@/components/LandingPage/Features/feature";
-import FooterSection from "@/components/LandingPage/FooterSection/FooterSection";
+import BusinessOutcomes from "@/components/LandingPage/BusinessSection";
+import ContactSection from "@/components/LandingPage/ContactSection";
+import FeaturesPage from "@/components/LandingPage/FeatureSection";
+import FooterSection from "@/components/LandingPage/FooterSection";
 import HeroSection from "@/components/LandingPage/HeroSection";
 import Navbar from "@/components/LandingPage/Navbar";
 import PricingSection from "@/components/LandingPage/PricingSection";
